location_lookup/requirements.txt
geopandas
sqlalchemy
pymysql
python-dotenv
